Fix github issue 182: Error when changing from roaming to local (1st time use).

+++ src/XmlNotepad/SettingsLoader.cs
@@ -140,7 +140,10 @@ namespace XmlNotepad
                     {
                         File.Delete(newLocation);
                     }
-                    File.Move(existingLocation, newLocation);
+                    if (File.Exists(existingLocation))
+                    {
+                        File.Move(existingLocation, newLocation);
+                    }                    
                     settings.FileName = newLocation;
                     this._settingsLocation = location;
                 }
